# move 2d fabric (low-latency) packet header updates to sender channel