--- a/tt_metal/fabric/hw/inc/edm_fabric/fabric_edm_packet_transmission.hpp
+++ b/tt_metal/fabric/hw/inc/edm_fabric/fabric_edm_packet_transmission.hpp
@@ -246,10 +246,10 @@ FORCE_INLINE void update_packet_header_for_next_hop(
 FORCE_INLINE void update_packet_header_for_next_hop(
     volatile tt_l1_ptr tt::tt_fabric::LowLatencyMeshPacketHeader* packet_header,
     tt::tt_fabric::LowLatencyMeshRoutingFields cached_routing_fields) {
-    // This is the hop index. At every ethernet hop, we increment by 1
-    // so that the next receiver indexes into its respecive hop command
-    // in packet_header.route_buffer[]
-    packet_header->routing_fields.value = cached_routing_fields.value + 1;
+    if constexpr (UPDATE_PKT_HDR_ON_RX_CH) {
+        packet_header->routing_fields.value = cached_routing_fields.value + 1;
+    }
+    // Intentionally empty - mesh routing updates the header on sender channel side
 }
 
 template <uint8_t NUM_SENDER_BUFFERS>
@@ -264,16 +264,20 @@ void update_packet_header_for_next_hop(
         tt::tt_fabric::edm_to_downstream_noc);
     std::uintptr_t offset =
         reinterpret_cast<std::uintptr_t>(&(packet_base->mcast_params[tt::tt_fabric::eth_chan_directions::EAST]));
-#else
-    tt::tt_fabric::LowLatencyMeshPacketHeader* packet_base = nullptr;
-    std::uintptr_t offset = reinterpret_cast<std::uintptr_t>(&(packet_base->routing_fields));
-#endif
     downstream_edm_interface.template update_edm_buffer_slot_word(offset, value, tt::tt_fabric::edm_to_downstream_noc);
+#else
+    if constexpr (UPDATE_PKT_HDR_ON_RX_CH) {
+        tt::tt_fabric::LowLatencyMeshPacketHeader* packet_base = nullptr;
+        std::uintptr_t offset = reinterpret_cast<std::uintptr_t>(&(packet_base->routing_fields));
+        downstream_edm_interface.template update_edm_buffer_slot_word(
+            offset, value, tt::tt_fabric::edm_to_downstream_noc);
+    }
+#endif
 }
 
 FORCE_INLINE void update_packet_header_for_next_hop(
-    volatile tt_l1_ptr tt::tt_fabric::MeshPacketHeader* packet_header,
-    tt::tt_fabric::LowLatencyMeshRoutingFields cached_routing_fields) {}
+    volatile tt_l1_ptr tt::tt_fabric::MeshPacketHeader* /*packet_header*/,
+    tt::tt_fabric::LowLatencyMeshRoutingFields /*cached_routing_fields*/) {}
 
 // This function forwards a packet to the downstream EDM channel for eventual sending
 // to the next chip in the line/ring