… tests/tt_metal/tt_metal/perf_microbenchmark/routing/golden/golden_bandwidth_summary_wormhole_b0_galaxy.csv
… tests/tt_metal/tt_metal/perf_microbenchmark/routing/golden/golden_bandwidth_summary_wormhole_b0_t3k.csv